Create optimizedrouter_v1.1.py
Router program to record the best and efficient routes and export the data to another program, you can implement a logging mechanism to keep track of routing decisions. In this example, I'll use a simple Python dictionary to record the routes and then export the data to a JSON file, which you can later import into the neural network model  I wrote earlier.